requirements.txt
discord.py>=1.7.0,<1.8.0
uvloop>=0.15.0,<0.16.0
python-dotenv
